Updated delete logic to add a confirmation dialogue

+++ app/src/main/java/com/example/passoff/itemView.kt
@@ -1,9 +1,11 @@
 package com.example.passoff
 
+import android.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.text.method.ScrollingMovementMethod
 import android.util.Log
+import android.view.LayoutInflater
 import android.widget.Button
 import android.widget.TextView
 import org.w3c.dom.Text
@@ -34,19 +36,30 @@ class itemView : AppCompatActivity() {
 
         val deleteButton = findViewById<Button>(R.id.delete_button)
         deleteButton.setOnClickListener {
-            val dbHandler = DBHandler(this)
-            val success = dbHandler.deletePassword(this.intent.getIntExtra("id", -1))
-            val loadDataTask = LoadDataTask(this)
-            loadDataTask.setRecyclerView(MainActivity.recyclerView)
-            loadDataTask.setCircularProgressIndicator(MainActivity.progressIndicator)
-            loadDataTask.execute()
-            if (success) {
-                Log.d("Database", "Password deleted successfully")
-            } else {
-                Log.e("Database", "Failed to delete password")
-            }
-            finish()
+            val dialogView = LayoutInflater.from(this).inflate(R.layout.dialog_confirm, null)
+            val dialog = AlertDialog.Builder(this)
+                .setTitle("Delete Password")
+                .setView(dialogView)
+                .setPositiveButton("Delete") { dialog, _ ->
+                    val dbHandler = DBHandler(this)
+                    val success = dbHandler.deletePassword(this.intent.getIntExtra("id", -1))
+                    if (success) {
+                        Log.d("Database", "Password deleted successfully")
+                    } else {
+                        Log.e("Database", "Failed to delete password")
+                    }
+                    val loadDataTask = LoadDataTask(this)
+                    loadDataTask.setRecyclerView(MainActivity.recyclerView)
+                    loadDataTask.setCircularProgressIndicator(MainActivity.progressIndicator)
+                    loadDataTask.execute()
+                    dialog.dismiss()
+                    finish()
+                }
+                .setNegativeButton("Cancel") { dialog, _ ->
+                    dialog.cancel()
+                }
+                .create()
+            dialog.show()
         }
-
     }
 }
